Refactor PDB identifier types from int to long
Changed various instances of PDB identifier fields and related logic from int to long across both frontend and backend components for consistency. Updated GraphQL queries, service methods, and relevant tests to accommodate the revised data type.

+++ backend/src/main/java/com/snazzyrobot/molviewbackend/controller/PdbDataController.java
@@ -28,12 +28,12 @@ public class PdbDataController {
     }
 
     @QueryMapping(name = "findPdbDataById")
-    public Optional<PdbData> findPdbDataById(@Argument int id) {
+    public Optional<PdbData> findPdbDataById(@Argument long id) {
         return pdbDataService.getPdbDataById(id);
     }
 
     @MutationMapping(name = "deletePdbData")
-    public Optional<PdbData> deletePdbData(@Argument int id) {
+    public Optional<PdbData> deletePdbData(@Argument long id) {
         return pdbDataService.deletePdbData(id);
     }
 }
