+++ backend/src/main/java/com/snazzyrobot/molviewbackend/repository/PdbDataRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PdbDataRepository extends JpaRepository<PdbData, Integer> {
+public interface PdbDataRepository extends JpaRepository<PdbData, Long> {
 }
